stripe_project/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
stripe==12.4.0
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
